+++ docs/CHANGELOG.md
@@ -8,11 +8,12 @@
 * The Progress page now shows progress towards reset.
 * Added some sources to the search dialog.
   * source:
-    * edz, titan, nessus, io, mercury, mars
+    * edz, titan, nessus, io, mercury, mars, tangled, dreaming
     * crucible, trials, ironbanner
-    * zavala, ikora, gunsmith, gambit, eververse
+    * zavala, ikora, gunsmith, gambit, eververse, shipwright
     * nm, do, fwc
-    * leviathan, lastwish
+    * leviathan, lastwish, sos, eow, prestige, raid
+    * prophecy, nightfall, adventure
 * In Chrome you can now Install DIM from the hamburger menu and use it as a standalone app. Chrome will support macOS later.
 
 # 4.76.0 (2018-11-04)